Labs/DVC_Exercise_Tracker/requirements.txt
dvc[gs]==3.55.2
pandas==2.2.3
scikit-learn==1.5.2
pyyaml==6.0.2
